Changes to be committed: 	modified:   docker-compose.yaml 	new file:   fastApiEL/migrations/versions/e0edb616b690_chande_auto.py 	modified:   fastApiEL/src/db/dals.py 	modified:   fastApiEL/src/db/models.py

+++ fastApiEL/src/db/models.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import BIGINT, Column, VARCHAR
+from sqlalchemy import Column, VARCHAR, BigInteger
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import declarative_base
 
@@ -11,8 +11,8 @@ class User(Base):
     __tablename__ = "users"
 
     id = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    id_tg = Column(BIGINT, nullable=False, unique=True)
-    education_id = Column(BIGINT, nullable=True)
-    group_id = Column(BIGINT, nullable=True)
+    id_tg = Column(BigInteger, nullable=False, unique=True)
+    education_id = Column(BigInteger, nullable=True)
+    group_id = Column(BigInteger, nullable=True)
     jwt_token = Column(VARCHAR, nullable=True)
